Added location service, comments, the ability to edit a profile.

+++ lib/firestor.dart
@@ -1,9 +1,10 @@
 import 'dart:io';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_storage/firebase_storage.dart';
+import 'package:flutter/foundation.dart';
 import 'package:foodgram/util/exception.dart';
 import 'model/user_model.dart';
-import 'package:uuid/uuid.dart';
 import 'storage.dart';
 
 
@@ -50,6 +51,27 @@ class Firebase_Firestor {
     }
   }
 
+  Future<void> updateUserProfile({
+    required String uid,
+    String? newUsername,
+    String? newProfilePictureUrl,
+  }) async {
+    try {
+      final userDoc = _firebaseFirestore.collection('users').doc(uid);
+
+      if (newUsername != null) {
+        await userDoc.update({'username': newUsername});
+      }
+
+      if (newProfilePictureUrl != null) {
+        await userDoc.update({'profile': newProfilePictureUrl});
+      }
+    } catch (e) {
+      throw Exception('Failed to update user profile: $e');
+    }
+  }
+
+
   Future<bool> CreatePost({
     required File postImage,
     required String caption,
@@ -71,6 +93,37 @@ class Firebase_Firestor {
       'time': data
     });
     return true;
+  }
+
+  Future<void> UpdatePost({
+    required String userId,
+    required String newUsername,
+    required String newProfilePictureUrl,
+  }) async {
+    try {
+      // Query for all posts by the user
+      final postsQuery = await _firebaseFirestore
+          .collection('posts')
+          .where('uid', isEqualTo: userId)
+          .get();
+
+      // Update each post with the new profile information
+      final batch = _firebaseFirestore.batch();
+      postsQuery.docs.forEach((postDoc) {
+        batch.update(postDoc.reference, {
+          'username': newUsername,
+          'profileImage': newProfilePictureUrl,
+        });
+      });
+
+      // Commit the batch update
+      await batch.commit();
+    } catch (e) {
+      if (kDebugMode) {
+        print('Error updating user posts: $e');
+      }
+      throw e; // Optionally, re-throw the error to handle it in the caller
+    }
   }
 
 
@@ -99,8 +152,24 @@ class Firebase_Firestor {
   Future<String> deletePost(String postId) async {
     String res = "Some error occurred";
     try {
-      await _firebaseFirestore.collection('posts').doc(postId).delete();
-      res = 'success';
+      var postDoc = await _firebaseFirestore.collection('posts').doc(postId).get();
+      if (postDoc.exists) {
+        // Get the photoImage URL from the post document
+        String postImageUrl = postDoc.data()?['postImage'];
+
+        // Delete the post document
+        await _firebaseFirestore.collection('posts').doc(postId).delete();
+
+        // Delete the image from Firebase Storage
+        if (postImageUrl != null) {
+          await FirebaseStorage.instance.refFromURL(postImageUrl).delete();
+        }
+        res = 'success';
+      }else{
+          res = 'success';
+
+        }
+
     } catch (err) {
       res = err.toString();
     }
@@ -133,6 +202,38 @@ class Firebase_Firestor {
     } catch (e) {
        print(e.toString());
     }
+  }
+
+  // Post comment
+  Future<String> postComment(String postId, String text, String uid,
+      String name, String profilePic) async {
+    String res = "Some error occurred";
+    try {
+      if (text.isNotEmpty) {
+        // if the likes list contains the user uid, we need to remove it
+        var docRef = await _firebaseFirestore.collection('posts') .doc(postId)
+            .collection('comments').doc();
+        String commentId = docRef.id;
+        _firebaseFirestore.collection('posts')
+            .doc(postId)
+            .collection('comments')
+            .doc(commentId)
+            .set({
+          'profilePic': profilePic,
+          'name': name,
+          'uid': uid,
+          'text': text,
+          'commentId': commentId,
+          'datePublished': DateTime.now(),
+        });
+        res = 'success';
+      } else {
+        res = "Please enter text";
+      }
+    } catch (err) {
+      res = err.toString();
+    }
+    return res;
   }
 
 }
